Final split of security group functions from utils
Also renamed _get_port_info_by_name -> get_port_info_by_name and
_get_ports -> get_ports.
